bug: fix sudo not working.
This was caused by $combine-maps being broken, since it reordered stuff
in $PATH and maps not preserving order. Rewrote with arrays and manual
deduplication, which isn't great. The cause of the breakage was that
sudo is in /run/wrappers/bin *and* /run/current-system/sw/bin, and map
reordered the unwrapped version (/run/current-system) above the wrapped
version (/run/wrappers).