fix: eww: music dropdown styling. Also: keyboard focus is off.
Yeah, this should of been atleast two commits, fuck off.

The layout is fairly sane now, although there are some issues still
with the current dropdown it's for all intents and purposes complete
enough for me to actually use it.

The issues that remain (which I may not deal with) are the following:

1. The menu updates slowly after a track changes due to the long poll
timer

2. Long titles grow the menu, and the following song will shrink them
again. It would be pretty slick if it truncated and rotated the title
like a radio or something, but that's a lot of work.

3. I played around with maybe adding a slider for seeking through the
song, but that was a bad idea since it looks like the implementation
for apps or playerctl is broken. I don't know which, and I don't really
care, but it doesn't work.